🚧 [100] Connected transactions to front

+++ src/app/components/money/accounts-list/accounts-list.ts
@@ -1,5 +1,6 @@
 import { Component, computed } from '@angular/core';
 import { MoneyService } from '@app/services/money.service';
+import { DefaultModal } from '@app/shared/components/default-modal/default-modal';
 import { Account, AccountKind } from '@app/shared/types';
 import { VButton } from '@ui-kit/components/v-button/v-button';
 import { VCard } from '@ui-kit/components/v-card/v-card';
@@ -9,13 +10,15 @@ import { AccountForm } from './account-form/account-form';
   selector: 'accounts-list',
   templateUrl: './accounts-list.html',
   standalone: true,
-  imports: [AccountForm, VButton, VCard],
+  imports: [AccountForm, DefaultModal, VButton, VCard],
 })
 export class AccountsList {
   protected accounts$$ = computed(() => this.moneyService.accounts$$());
   protected currencies$$ = computed(() => this.moneyService.currencies$$());
   protected showForm = false;
   protected editingAccount: Account | null = null;
+  protected isDeleteConfirmOpen = false;
+  protected pendingDeleteId: number | null = null;
 
   constructor(private moneyService: MoneyService) {}
 
@@ -30,6 +33,24 @@ export class AccountsList {
   }
 
   protected deleteAccount(id: number): void {
+    this.openConfirmationModal(id);
+  }
+
+  protected openConfirmationModal(id: number): void {
+    this.pendingDeleteId = id;
+    this.isDeleteConfirmOpen = true;
+  }
+
+  protected closeConfirmationModal(): void {
+    this.isDeleteConfirmOpen = false;
+    this.pendingDeleteId = null;
+  }
+
+  protected onDeleteConfirmed(): void {
+    const id = this.pendingDeleteId;
+    this.isDeleteConfirmOpen = false;
+    this.pendingDeleteId = null;
+    if (!id) return;
     this.moneyService.deleteAccount(id).subscribe((success) => {});
   }
